PR response (combined recorder with series block frame)
