+++ src/dbnode/encoding/tile/frame_units_test.go
@@ -24,6 +24,7 @@ import (
 	"testing"
 
 	xtime "github.com/m3db/m3/src/x/time"
+
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
