+++ src/dbnode/encoding/tile/series_block_frame_test.go
@@ -26,6 +26,7 @@ import (
 
 	"github.com/m3db/m3/src/dbnode/ts"
 	xtime "github.com/m3db/m3/src/x/time"
+
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
